# worker/openapi.json
{
  "components": {
    "schemas": {
      "BlobInfo": {
        "properties": {
          "key": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          }
        },
        "required": [
          "key",
          "size"
        ],
        "type": "object"
      },
      "Error": {
        "$ref": "#/components/schemas/Error"
      },
...
      "JobCreated": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobView"
          },
          "upload": {
            "type": "string"
          }
        },
        "required": [
          "job",
          "upload"
        ],
        "type": "object"
      },
...
      "JobList": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "type": "object"
      },
...
      "JobSpec": {
        "properties": {
          "args": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "gui": {
            "type": "boolean"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "timeout_s": {
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "name",
          "size",
          "sha256"
        ],
        "type": "object"
      },
      "JobView": {
        "properties": {
          "cancel_requested": {
            "type": "boolean"
          },
          "created": {
            "format": "date-time",
            "type": "string"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/BlobInfo"
            },
            "type": "array"
          },
          "finished": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lease": {
            "format": "date-time",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "outcome": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "position": {
            "type": "integer"
          },
          "queued": {
            "format": "date-time",
            "type": "string"
          },
          "runner": {
            "type": "string"
          },
          "running": {
            "type": "integer"
          },
          "spec": {
            "$ref": "#/components/schemas/JobSpec"
          },
          "started": {
            "format": "date-time",
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "why": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner",
          "spec",
          "state",
          "created",
          "queued",
          "started",
          "finished",
          "lease",
          "running"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "glaze-status-write": {
        "description": "CI's conformance job posting a glaze run. The Worker's secret GLAZE_STATUS_TOKEN; clients read it from GLAZE_STATUS_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "golden-read": {
        "description": "reading the golden image: every machine that pulls. The Worker's secret GOLDEN_TOKEN; clients read it from IRGO_GOLDEN_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "golden-write": {
        "description": "writing, deleting and listing it: only where you push. The Worker's secret GOLDEN_PUSH_TOKEN; clients read it from IRGO_GOLDEN_PUSH_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "jobs": {
        "description": "a remote job's caller: submits and reads its own jobs (the secret is name=token,name=token, one per caller). The Worker's secret JOBS_TOKENS; clients read it from IRGO_REMOTE_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "jobs-admin": {
        "description": "lists every job in the queue. The Worker's secret JOBS_ADMIN_TOKEN; clients read it from IRGO_REMOTE_ADMIN_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "jobs-runner": {
        "description": "the Mac running irgo-winvm serve: takes jobs and reports them. The Worker's secret JOBS_RUNNER_TOKEN; clients read it from IRGO_REMOTE_RUNNER_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "ledger-read": {
        "description": "reading the ledger: its page and its JSON. The Worker's secret LEDGER_READ_TOKEN; clients read it from IRGO_LEDGER_READ_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      },
      "ledger-write": {
        "description": "the tool posting events to the ledger. The Worker's secret LEDGER_TOKEN; clients read it from IRGO_LEDGER_TOKEN.",
        "scheme": "bearer",
        "type": "http"
      }
...
  },
  "info": {
...
    "title": "irgo-windows-vm Worker",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs": {
      "get": {
        "operationId": "job-list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code internal"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code storage"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code not-configured"
          }
        },
        "security": [
          {
            "jobs-admin": []
          }
        ],
        "summary": "every job in the index, live and ended in the last day",
        "x-irgo-winvm-commands": [
          "remote-status"
        ],
        "x-scope": "jobs-admin"
      },
      "post": {
        "operationId": "job-submit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCreated"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code bad-request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code unauthorized"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code too-large"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code too-many"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code internal"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code storage"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "code busy or not-configured"
          }
        },
        "security": [
          {
            "jobs": []
          }
        ],
        "summary": "submit a job spec; the job waits for its binary at the answer's upload path",
        "x-irgo-winvm-commands": [
          "remote-submit"
        ],
        "x-max-body": 65536,
        "x-scope": "jobs"
      }
    },
...
  }
}